Added a simplified interface to sheet->table
This took a bit of work. There's now a "guess-sanitizer", which allows a list of
data to be guessed at.

This makes it easier to write

(sheet->table "Foo" "http://...")

and let the library guess what each column is, instead of requiring the user
to specify sanitizers for every column in the spreadsheet.