Problematic sql-null...
If importing from MySQL, it is possible for any column, regardless of type,
that you may have an #<sql-null> value.

This should probably not kill the import, because there's no way to clean the data
further before using the (read-sql ...) interface.

Does this mean the sanitizers allow unsanitary data through? If so, what are they?

I've added this, so now the (for example) integer-sanitizer allows only integers through,
as well as sql-null.

Elch.