Documentation, SQL cleanup
I've broken SQL apart, and it is a bit more modular. All SQL formats should now be a few data tables for
conversions, and beyond that, common code. This means that adding some improved error handling, etc., will
be a cost paid that applies to all of the different SQL database types, and not on a per-db basis.

The documentation is improving. Broke it apart into a Reference and Guide. The reference is just
documentation about the functions themselves, and the Guide is more "how to use it."