Undid changes to add-series.
Although it would be nice for it to be fully functional, that would imply a very different structure underneath. For example, I can't have shared structures burried in the table; my simplistic transformation would have left all the gvectors shared without a deep copy.

Which, for adding a series to a table, seems a bit much.

I do need a deep copy, though... and, I wonder if I should always return a clean table/copy.

It feels like the right thing to do.